+++ fud2/fud-core/src/exec/plan/mod.rs
@@ -1,3 +1,7 @@
+mod enumerative_planner;
 mod planner;
+mod single_op_output_planner;
 
-pub use planner::{EnumeratePlanner, FindPlan, SingleOpOutputPlanner};
+pub use enumerative_planner::EnumeratePlanner;
+pub use planner::{FindPlan, Step};
+pub use single_op_output_planner::SingleOpOutputPlanner;
